backend/requirements.txt
setuptools>=72.0
fastapi==0.115.6
uvicorn==0.34.0
aiosqlite==0.20.0
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
openai-whisper
piper-tts
harperdb
